refactor(): more tests and stability

+++ bin/build-module-map.sh
@@ -3,7 +3,8 @@
 declare root="$(dirname "$(dirname "$(realpath "${BASH_SOURCE[0]}")")")"
 
 mkdir -p "$root/build/modules"
-rm -f "$root/build/modules/modules.modulemap"
-rm -rf "$root/build/cache"
-cp "$root/src/modules.modulemap" "$root/build/modules"
-echo " info: build modules.modulemap"
+if ! test -f "$root/build/modules/modules.modulemap" || (( $(stat "$root/src/modules.modulemap" -c %Y) > $(stat "$root/build/modules/modules.modulemap" -c %Y) )); then
+  rm -f "$root/build/modules/modules.modulemap"
+  cp "$root/src/modules.modulemap" "$root/build/modules"
+  echo " info: build modules.modulemap"
+fi
